feat(core)!: define OutputFlags for side-chain contracts (#4088)
Description
---
Defines OutputFlags for side-chain contracts.

Motivation and Context
---
Define flags required for side-chain contracts. Deprecated flags are kept and can be removed in implementation PRs
for the various side-chain components.

How Has This Been Tested?
---
Unit and cucumber tests updated as necessary

+++ base_layer/core/src/proto/transaction.rs
@@ -282,7 +282,7 @@ impl TryFrom<proto::types::OutputFeatures> for OutputFeatures {
             Some(PublicKey::from_bytes(features.parent_public_key.as_bytes()).map_err(|err| format!("{:?}", err))?)
         };
 
-        let flags = u8::try_from(features.flags).map_err(|_| "Invalid output flags: overflowed u8")?;
+        let flags = u16::try_from(features.flags).map_err(|_| "Invalid output flags: overflowed u8")?;
 
         Ok(OutputFeatures::new(
             OutputFeaturesVersion::try_from(
